+++ main/include/maze_solver.hpp
@@ -7,10 +7,10 @@
 
 using namespace std;
 constexpr unsigned char MAX_MAZE_SIZE = 32;
-constexpr double ROOT2 = sqrt(2);
-constexpr double dt = 0.001;
-constexpr double eps = 0.001;
-constexpr double PI = 3.141592653589793238;
+constexpr float ROOT2 = sqrt(2);
+constexpr float dt = 0.001;
+constexpr float eps = 0.001;
+constexpr float PI = 3.141592653589793238;
 
 typedef struct {
   float n;
